+++ CHANGELOG.md
@@ -9,3 +9,13 @@ Fail2ban admin panel in flask app
 ### Added
     - Api rest for commands fail2ban
     - Initial template
+
+## [1.0.2] - 2023-03-27
+
+### Added
+    - requirements.txt
+    - module requests
+    - flash messages
+    - fail2ban status
+    - fail2ban list bans
+    - fail2ban add ban and remove ban